Added jest test cases and json server

+++ src/app/quote/quotes-list/quotes-list.component.ts
@@ -1,11 +1,13 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { Quote } from '../models/quote';
-import { Observable } from 'rxjs';
-import { QuotesService } from '../services/quotes.service';
-import { Store } from '@ngrx/store';
-import { AppState, quoteState } from '../store/app.state';
-import { GetQuotes, DeleteQuotes } from '../quote-store/actions/quote.actions';
+import { Component, Input, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
+import { Store } from '@ngrx/store';
+import { Observable } from 'rxjs';
+import { Quote } from '../../models/quote';
+import {
+  DeleteQuotes,
+  GetQuotes,
+} from '../../quote-store/actions/quote.actions';
+import { AppState, quoteState } from '../../store/app.state';
 
 @Component({
   selector: 'app-quotes-list',
@@ -14,10 +16,15 @@ import { Router } from '@angular/router';
 })
 export class QuotesListComponent implements OnInit {
   quotes: Quote[];
+
   createQuote: boolean;
+
   @Input() userAuthenticated: boolean;
+
   displayMessage: string;
+
   quoteToUpdate: Quote;
+
   getState: Observable<any>;
 
   constructor(private router: Router, private store: Store<AppState>) {
@@ -37,28 +44,29 @@ export class QuotesListComponent implements OnInit {
   }
 
   public deleteQuote(id: number): void {
-    this.store.dispatch(new DeleteQuotes(id));
-    this.store.dispatch(new GetQuotes());
+    const deleteResult = window.confirm(
+      'Are you sure you want to delete this quote?'
+    );
+    if (deleteResult) {
+      this.store.dispatch(new DeleteQuotes(id));
+      this.store.dispatch(new GetQuotes());
+    }
   }
 
-  public quoteCreated(event): void {
+  public quoteCreated(event: any): void {
     this.displayMessage = event;
     this.loadAllQuotes();
   }
 
-  public quoteNotCreated(event): void {
+  public quoteNotCreated(event: any): void {
     this.displayMessage = event;
   }
 
   public createQuoteForm(): void {
-    // localStorage.removeItem('quote');
     this.router.navigateByUrl('create');
   }
 
   public updateQuote(quoteToUpdate: Quote): void {
-    // localStorage.removeItem('quote');
-    // localStorage.setItem('quote', JSON.stringify(quoteToUpdate));
-    // this.router.navigate(['update', { id: quoteToUpdate.id }]);
     this.router.navigateByUrl('update/' + quoteToUpdate.id);
   }
